requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dacite==1.8.1
deap==1.4.1
debugpy==1.8.5
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.20.0
fonttools==4.53.1
fpdf==1.7.2
htmlmin==0.1.12
idna==3.7
ImageHash==4.3.1
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
kneed==0.8.5
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
multimethod==1.12
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
phik==0.12.4
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pyswarm==0.6
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.6.0
pywin32==306
PyYAML==6.0.1
pyzmq==26.1.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
scikit-learn==1.5.1
scipy==1.11.4
seaborn==0.12.2
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.2
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visions==0.7.6
wcwidth==0.2.13
wordcloud==1.9.3
ydata-profiling==4.7.0